+++ ci/cpplint.sh
@@ -23,24 +23,12 @@ then
   CXX=g++
 fi
 
-# CFLAGS need to be set for some libraries that require a special include path
-# for their headers.
-pkg-config --cflags jsoncpp
-if [[ $? -eq 0 ]]
-then
-  CFLAGS=$(pkg-config --cflags jsoncpp)
-else
-  # just use some standard path for libjsoncpp includes
-  echo "Info: Falling back to standard includes for jsoncpp!"
-  CFLAGS="-I/usr/include/jsoncpp"
-fi
-
 # print $CXX version to see which version is used for syntax check
 $CXX --version
 echo
 
 # find all C++ files and run them through the compilers's syntax check
-find ./ \( -name '*.cpp' -o -name '*.hpp' -o -name '*.h' \) -print0 | xargs -0 -i $CXX $CFLAGS -fsyntax-only -Wall -std=c++17 -Dwic_owm_find_location -Dwic_weather_comparison {}
+find ./ \( -name '*.cpp' -o -name '*.hpp' -o -name '*.h' \) -print0 | xargs -0 -i $CXX -fsyntax-only -Wall -std=c++17 -Dwic_owm_find_location -Dwic_weather_comparison {}
 if [[ $? -ne 0 ]]
 then
   echo "Some source code files contain syntax errors!"